relocating cloud functions folder (fix codeclimate integration issue)
