+++ .env.example.dev
@@ -1,0 +1,12 @@
+ENV=dev
+FSTORE_EMULATOR=true
+
+TYPEORM_CONNECTION=mysql
+TYPEORM_HOST=localhost
+TYPEORM_PORT=3306
+TYPEORM_USERNAME=root
+TYPEORM_PASSWORD=development
+TYPEORM_DATABASE=dev_x
+TYPEORM_SYNCHRONIZE=true
+TYPEORM_LOGGING=false
+TYPEORM_MIGRATIONS_DIR=src/database/migrations
