--- functions/.env.example.dev
@@ -1,9 +1,0 @@
-# Environment
-ENV=dev
-
-# Rate Limit values
-RATE_LIMIT_WINDOW=5
-RATE_LIMIT_MAX_REQUESTS=100
-
-# Firestore Emulator
-FSTORE_EMULATOR=true
