--- functions/.gitignore
@@ -1,9 +1,0 @@
-# Compiled JavaScript files
-dist
-
-# Node.js dependency directory
-node_modules/
-.env.dev
-.env.test
-.env.prod
-.nvmrc
