--- functions/tsconfig.dev.json
@@ -1,3 +1,0 @@
-{
-  "include": [".eslintrc.js", "jest.config.js"]
-}
