--- scripts/clear-all-data-firestore-emulator.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-curl -v -X DELETE "http://localhost:8080/emulator/v1/projects/quicktests-b86f7-2cc2b/databases/(default)/documents"
