+++ src/database/entities/collectionboard.entity.ts
