+++ src/database/entities/financialtransaction.entity.ts
