+++ src/database/seeds/entities/financialtransaction.seed.ts
